+++ clients/derivatives-trading-portfolio-margin-pro/examples/rest-api/AccountApi/changeAutoRepayFuturesStatus.ts
@@ -13,7 +13,7 @@ const client = new DerivativesTradingPortfolioMarginPro({ configurationRestAPI }
 async function changeAutoRepayFuturesStatus() {
     try {
         const response = await client.restAPI.changeAutoRepayFuturesStatus({
-            autoRepay: '',
+            autoRepay: 'true',
         });
 
         const rateLimits = response.rateLimits!;