+++ clients/derivatives-trading-portfolio-margin-pro/src/rest-api/types/mint-bfusd-for-portfolio-margin-response.ts
@@ -48,5 +48,5 @@ export interface MintBfusdForPortfolioMarginResponse {
      * @type {number}
      * @memberof MintBfusdForPortfolioMarginResponse
      */
-    rate?: number;
+    mintRate?: number;
 }