+++ clients/derivatives-trading-portfolio-margin-pro/src/rest-api/types/redeem-bfusd-for-portfolio-margin-response.ts
@@ -48,5 +48,5 @@ export interface RedeemBfusdForPortfolioMarginResponse {
      * @type {number}
      * @memberof RedeemBfusdForPortfolioMarginResponse
      */
-    rate?: number;
+    redeemRate?: number;
 }